+++ src/Eccube/Service/PurchaseFlow/Processor/DeliveryFeePreprocessor.php
@@ -14,6 +14,7 @@
 namespace Eccube\Service\PurchaseFlow\Processor;
 
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\NoResultException;
 use Eccube\Entity\BaseInfo;
 use Eccube\Entity\DeliveryFee;
 use Eccube\Entity\ItemHolderInterface;
@@ -34,75 +35,55 @@ use Eccube\Service\PurchaseFlow\PurchaseContext;
  */
 class DeliveryFeePreprocessor implements ItemHolderPreprocessor
 {
-    /** @var BaseInfo */
-    protected $BaseInfo;
-
-    /**
-     * @var EntityManagerInterface
-     */
-    protected $entityManager;
-
-    /**
-     * @var TaxRuleRepository
-     */
-    protected $taxRuleRepository;
-
-    /**
-     * @var DeliveryFeeRepository
-     */
-    protected $deliveryFeeRepository;
+    protected BaseInfo $BaseInfo;
 
     /**
      * DeliveryFeePreprocessor constructor.
-     *
-     * @param BaseInfoRepository $baseInfoRepository
-     * @param EntityManagerInterface $entityManager
-     * @param TaxRuleRepository $taxRuleRepository
-     * @param DeliveryFeeRepository $deliveryFeeRepository
      */
     public function __construct(
         BaseInfoRepository $baseInfoRepository,
-        EntityManagerInterface $entityManager,
-        TaxRuleRepository $taxRuleRepository,
-        DeliveryFeeRepository $deliveryFeeRepository,
+        protected EntityManagerInterface $entityManager,
+        protected TaxRuleRepository $taxRuleRepository,
+        protected DeliveryFeeRepository $deliveryFeeRepository,
     ) {
         $this->BaseInfo = $baseInfoRepository->get();
-        $this->entityManager = $entityManager;
-        $this->taxRuleRepository = $taxRuleRepository;
-        $this->deliveryFeeRepository = $deliveryFeeRepository;
     }
 
     /**
-     * @param ItemHolderInterface $itemHolder
-     * @param PurchaseContext $context
+     * @param ItemHolderInterface $itemHolder カート or 注文
+     * @param PurchaseContext $context 購入フローのコンテキスト
      *
-     * @throws \Doctrine\ORM\NoResultException
+     * @throws NoResultException
      */
-    public function process(ItemHolderInterface $itemHolder, PurchaseContext $context)
+    #[\Override]
+    public function process(ItemHolderInterface $itemHolder, PurchaseContext $context): void
     {
-        $this->removeDeliveryFeeItem($itemHolder);
-        $this->saveDeliveryFeeItem($itemHolder);
+        if ($itemHolder instanceof Order) {
+            $this->removeDeliveryFeeItem($itemHolder);
+            $this->saveDeliveryFeeItem($itemHolder);
+        }
     }
 
-    private function removeDeliveryFeeItem(ItemHolderInterface $itemHolder)
+    private function removeDeliveryFeeItem(ItemHolderInterface $itemHolder): void
     {
-        foreach ($itemHolder->getShippings() as $Shipping) {
-            foreach ($Shipping->getOrderItems() as $item) {
-                if ($item->getProcessorName() == DeliveryFeePreprocessor::class) {
-                    $Shipping->removeOrderItem($item);
-                    $itemHolder->removeOrderItem($item);
-                    $this->entityManager->remove($item);
+        if ($itemHolder instanceof Order) {
+            foreach ($itemHolder->getShippings() as $Shipping) {
+                /** @var OrderItem $item */
+                foreach ($Shipping->getOrderItems() as $item) {
+                    if ($item->getProcessorName() == DeliveryFeePreprocessor::class) {
+                        $Shipping->removeOrderItem($item);
+                        $itemHolder->removeOrderItem($item);
+                        $this->entityManager->remove($item);
+                    }
                 }
             }
         }
     }
 
     /**
-     * @param ItemHolderInterface $itemHolder
-     *
-     * @throws \Doctrine\ORM\NoResultException
+     * @throws NoResultException
      */
-    private function saveDeliveryFeeItem(ItemHolderInterface $itemHolder)
+    private function saveDeliveryFeeItem(ItemHolderInterface $itemHolder): void
     {
         $DeliveryFeeType = $this->entityManager
             ->find(OrderItemType::class, OrderItemType::DELIVERY_FEE);
@@ -116,14 +97,14 @@ class DeliveryFeePreprocessor implements ItemHolderPreprocessor
         /* @var Shipping $Shipping */
         foreach ($Order->getShippings() as $Shipping) {
             // 送料の計算
-            $deliveryFeeProduct = 0;
+            $deliveryFeeProduct = '0';
             if ($this->BaseInfo->isOptionProductDeliveryFee()) {
                 /** @var OrderItem $item */
                 foreach ($Shipping->getOrderItems() as $item) {
                     if (!$item->isProduct()) {
                         continue;
                     }
-                    $deliveryFeeProduct += $item->getProductClass()->getDeliveryFee() * $item->getQuantity();
+                    $deliveryFeeProduct = bcadd($deliveryFeeProduct, bcmul((string) $item->getProductClass()->getDeliveryFee(), (string) $item->getQuantity(), 0));
                 }
             }
 
@@ -132,15 +113,15 @@ class DeliveryFeePreprocessor implements ItemHolderPreprocessor
                 'Delivery' => $Shipping->getDelivery(),
                 'Pref' => $Shipping->getPref(),
             ]);
-            $fee = is_object($DeliveryFee) ? $DeliveryFee->getFee() : 0;
+            $fee = is_object($DeliveryFee) ? $DeliveryFee->getFee() : '0';
 
             $OrderItem = new OrderItem();
             $OrderItem->setProductName($DeliveryFeeType->getName())
-                ->setPrice($fee + $deliveryFeeProduct)
-                ->setQuantity(1)
+                ->setPrice(bcadd($fee, $deliveryFeeProduct))
+                ->setQuantity('1')
                 ->setOrderItemType($DeliveryFeeType)
                 ->setShipping($Shipping)
-                ->setOrder($itemHolder)
+                ->setOrder($Order)
                 ->setTaxDisplayType($TaxInclude)
                 ->setTaxType($Taxation)
                 ->setProcessorName(DeliveryFeePreprocessor::class);
